fix voiceCard audio playing bug
